+++ lib/hovr_app_update.dart
@@ -2,4 +2,7 @@ library;
 
 export 'src/app_info.dart';
 export 'src/app_update_config.dart';
+export 'src/app_update_result.dart';
 export 'src/hovr_app_update.dart';
+export 'src/ota/ota_update_status.dart';
+export 'src/ota/ota_updater.dart';